requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
build==1.2.2
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
cryptography==43.0.1
decorator==4.4.2
deprecation==2.1.0
dj-database-url==2.2.0
Django==5.1
django-jazzmin==3.0.0
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==28.1.0
ffmpeg-python==0.2.0
firebase-admin==6.5.0
frozenlist==1.5.0
future==1.0.0
gcloud==0.17.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
googlemaps==4.10.0
gotrue==2.10.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
imageio==2.36.0
imageio-ffmpeg==0.5.1
jws==0.1.3
kombu==5.4.2
moviepy==1.0.3
msgpack==1.0.8
multidict==6.1.0
mysqlclient==2.2.4
numpy==2.1.3
oauth2client==3.0.0
oauthlib==3.2.2
packaging==24.1
Pillow==9.5.0
pip-tools==7.4.1
pipdeptree==2.23.3
postgrest==0.18.0
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.4.3
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.1.4
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-jwt==2.0.1
realtime==2.0.6
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.7.0
rest-framework-simplejwt==0.0.2
rsa==4.9
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tomli==2.0.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websockets==13.1
wheel==0.44.0
yarl==1.17.1
